Remove alignments from the root level of the site editor
